Move MsalProvider to wrap AuthContext
According to the @azure/msal-react migration guide, interactive methods
such as loginPopup should only be called inside the context of
MsalProvider.
