+++ src/GeoCop.Frontend/src/App.jsx
@@ -1,4 +1,6 @@
-import { useEffect, useState } from "react";
+import { PublicClientApplication } from "@azure/msal-browser";
+import { MsalProvider } from "@azure/msal-react";
+import { useEffect, useMemo, useState } from "react";
 import { Alert } from "react-bootstrap";
 import { BrowserRouter as Router, Routes, Route, Navigate } from "react-router-dom";
 import { Snackbar } from "@mui/material";
@@ -86,65 +88,72 @@ export const App = () => {
   const openModalContent = (content, type) =>
     setModalContent(content) & setModalContentType(type) & setShowModalContent(true);
 
+  const oauth = clientSettings?.oauth;
+  const msalInstance = useMemo(() => {
+    return new PublicClientApplication(oauth ?? {});
+  }, [oauth]);
+
   return (
-    <AuthProvider authScopes={clientSettings?.authScopes} oauth={clientSettings?.oauth} onLoginError={setAlertText}>
-      <div className="app">
-        <Router>
-          <Header clientSettings={clientSettings} />
-          <Routes>
-            <Route
-              exact
-              path="/"
-              element={
-                <Home
-                  clientSettings={clientSettings}
-                  nutzungsbestimmungenAvailable={nutzungsbestimmungenContent ? true : false}
-                  showNutzungsbestimmungen={() => openModalContent(nutzungsbestimmungenContent, "markdown")}
-                  quickStartContent={quickStartContent}
-                  setShowBannerContent={setShowBannerContent}
-                />
-              }
-            />
-          </Routes>
-          <LoggedOutTemplate>
+    <MsalProvider instance={msalInstance}>
+      <AuthProvider authScopes={clientSettings?.authScopes} onLoginError={setAlertText}>
+        <div className="app">
+          <Router>
+            <Header clientSettings={clientSettings} />
             <Routes>
-              <Route path="*" element={<Navigate to="/" />} />
+              <Route
+                exact
+                path="/"
+                element={
+                  <Home
+                    clientSettings={clientSettings}
+                    nutzungsbestimmungenAvailable={nutzungsbestimmungenContent ? true : false}
+                    showNutzungsbestimmungen={() => openModalContent(nutzungsbestimmungenContent, "markdown")}
+                    quickStartContent={quickStartContent}
+                    setShowBannerContent={setShowBannerContent}
+                  />
+                }
+              />
             </Routes>
-          </LoggedOutTemplate>
-          <AdminTemplate>
-            <Routes>
-              <Route path="/admin" element={<Admin />} />
-            </Routes>
-          </AdminTemplate>
-        </Router>
-        <Footer
-          openModalContent={openModalContent}
-          infoHilfeContent={infoHilfeContent}
-          nutzungsbestimmungenContent={nutzungsbestimmungenContent}
-          datenschutzContent={datenschutzContent}
-          impressumContent={impressumContent}
-          clientSettings={clientSettings}
-          appVersion={backendVersion}
-          licenseInfoCustom={licenseInfoCustom}
-          licenseInfo={licenseInfo}
-        />
-        <ModalContent
-          className="modal"
-          show={showModalContent}
-          content={modalContent}
-          type={modalContentType}
-          onHide={() => setShowModalContent(false)}
-        />
-        {bannerContent && showBannerContent && (
-          <BannerContent className="banner" content={bannerContent} onHide={() => setShowBannerContent(false)} />
-        )}
-      </div>
-      <Snackbar key={alertText} open={alertText !== ""} anchorOrigin={{ vertical: "top", horizontal: "right" }}>
-        <Alert variant="danger" dismissible onClose={() => setAlertText("")}>
-          <p>{alertText}</p>
-        </Alert>
-      </Snackbar>
-    </AuthProvider>
+            <LoggedOutTemplate>
+              <Routes>
+                <Route path="*" element={<Navigate to="/" />} />
+              </Routes>
+            </LoggedOutTemplate>
+            <AdminTemplate>
+              <Routes>
+                <Route path="/admin" element={<Admin />} />
+              </Routes>
+            </AdminTemplate>
+          </Router>
+          <Footer
+            openModalContent={openModalContent}
+            infoHilfeContent={infoHilfeContent}
+            nutzungsbestimmungenContent={nutzungsbestimmungenContent}
+            datenschutzContent={datenschutzContent}
+            impressumContent={impressumContent}
+            clientSettings={clientSettings}
+            appVersion={backendVersion}
+            licenseInfoCustom={licenseInfoCustom}
+            licenseInfo={licenseInfo}
+          />
+          <ModalContent
+            className="modal"
+            show={showModalContent}
+            content={modalContent}
+            type={modalContentType}
+            onHide={() => setShowModalContent(false)}
+          />
+          {bannerContent && showBannerContent && (
+            <BannerContent className="banner" content={bannerContent} onHide={() => setShowBannerContent(false)} />
+          )}
+        </div>
+        <Snackbar key={alertText} open={alertText !== ""} anchorOrigin={{ vertical: "top", horizontal: "right" }}>
+          <Alert variant="danger" dismissible onClose={() => setAlertText("")}>
+            <p>{alertText}</p>
+          </Alert>
+        </Snackbar>
+      </AuthProvider>
+    </MsalProvider>
   );
 };
